move audio file under import domain
